# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: "Fluxa API"
  version: "1.0.0"
  description: |
    Fluxa is a multi-currency digital asset platform built on Stellar. It provides
    wallets, instant transfers, FX conversions, scheduled recurring payments,
    batch disbursements, fiat on/off-ramps, and a developer-friendly webhook system.

    All endpoints under `/v1` require a Bearer API key in the `Authorization` header.
  contact:
    name: "Fluxa Engineering"
    email: "engineering@fluxa.io"
    url: "https://github.com/fluxa/fluxa"
  license:
    name: "MIT"
    url: "https://opensource.org/licenses/MIT"
servers:
  - url: http://localhost:3000
    description: "Local development"
  - url: https://api.fluxa.io
    description: "Production"
...
security:
  - BearerAuth: []
paths:
...
  /v1/keys:
    post:
      operationId: createAPIKey
      summary: "Create an API key"
      description: |
        Generates a new API key for the authenticated tenant. The raw key is
        returned only in this response — it cannot be retrieved later.
      tags: ["Auth"]
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateKeyRequest"
            example:
              label: "CI deploy key"
      responses:
        "201":
          description: "API key created"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIKeyCreated"
              example:
                id: "k_9a8b7c6d"
                key: "sk_live_a1b2c3d4e5f6g7h8"
                prefix: "sk_live_a1b2"
                label: "CI deploy key"
                created_at: "2026-01-15T10:00:00Z"
        "401":
          $ref: "#/components/responses/Unauthorized"

    get:
      operationId: listAPIKeys
      summary: "List API keys"
      description: "Returns all API keys for the authenticated tenant."
      tags: ["Auth"]
      responses:
        "200":
          description: "List of API keys"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/APIKeySummary"
              example:
                - id: "k_9a8b7c6d"
                  prefix: "sk_live_a1b2"
                  label: "CI deploy key"
                  last_used_at: "2026-01-15T12:00:00Z"
                  revoked_at: null
                  created_at: "2026-01-15T10:00:00Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/keys/{id}:
    delete:
      operationId: revokeAPIKey
      summary: "Revoke an API key"
      description: "Permanently revokes the given API key. The key can no longer be used."
      tags: ["Auth"]
      parameters:
        - $ref: "#/components/parameters/KeyID"
      responses:
        "204":
          description: "Key revoked"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Wallets ───────────────────────────────────────────────────────────────
...
  /v1/transfers:
    post:
      operationId: initiateTransfer
      summary: "Initiate a transfer"
      description: "Moves an asset between two wallets, applying platform fees."
      tags: ["Transfers"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransferRequest"
            example:
              from_wallet_id: "w_1a2b3c4d"
              to_wallet_id: "w_5e6f7g8h"
              asset: "USDC"
              amount: "100.0000000"
      responses:
        "202":
          description: "Transfer accepted"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferResponse"
              example:
                id: "tx_abc123"
                tx_hash: ""
                type: "transfer"
                status: "pending"
                from_wallet_id: "w_1a2b3c4d"
                to_wallet_id: "w_5e6f7g8h"
                asset: "USDC"
                amount: "100.0000000"
                fee_amount: "0.2500000"
                net_amount: "99.7500000"
                fee_bps: 25
                created_at: "2026-01-15T10:00:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

    get:
      operationId: listTransfers
      summary: "List transfers"
      description: "Returns all transactions for a given wallet."
      tags: ["Transactions"]
      parameters:
        - name: wallet_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
          description: "Filter by wallet ID"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
          description: "Maximum number of results"
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
          description: "Offset for pagination"
      responses:
        "200":
          description: "List of transactions"
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactions:
                    type: array
                    items:
                      $ref: "#/components/schemas/TransferResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/transfers/{id}:
    get:
      operationId: getTransfer
      summary: "Get transfer details"
      description: "Returns full details for a single transfer."
      tags: ["Transfers"]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: "Transaction ID"
      responses:
        "200":
          description: "Transfer details"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Batch Transfers ───────────────────────────────────────────────────────
...
components:
...
  securitySchemes:
    BearerAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Pass a tenant API key as `Bearer <key>` in the `Authorization` header.
        Keys are prefixed with `sk_live_` (production) or `sk_test_` (sandbox).
    AdminAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Pass an admin API key as `Bearer <key>`. Admin keys are prefixed with
        `sk_admin_`.

  # ── Parameters ──────────────────────────────────────────────────────────
  parameters:
...
    KeyID:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: "API key ID"
...
  responses:
    BadRequest:
      description: "Bad request"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: "BAD_REQUEST"
              message: "invalid request body"
    Unauthorized:
      description: "Unauthorized — missing or invalid API key"
      content:
        application/json:
          schema:
            type: object
          example:
            error: "missing or invalid authorization header"
    NotFound:
      description: "Resource not found"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: "NOT_FOUND"
              message: "resource not found"
...
  schemas:
...
    ErrorResponse:
      type: object
      required: ["error"]
      properties:
        error:
          type: object
          required: ["code", "message"]
          properties:
            code:
              type: string
              enum:
                - "BAD_REQUEST"
                - "NOT_FOUND"
                - "UNAUTHORIZED"
                - "FORBIDDEN"
                - "INTERNAL_ERROR"
                - "QUOTE_EXPIRED"
                - "QUOTE_ALREADY_USED"
              description: "Machine-readable error code"
            message:
              type: string
              description: "Human-readable error message"

    # ── Health ────────────────────────────────────────────────────────────
...
    CreateKeyRequest:
      type: object
      properties:
        label:
          type: string
          nullable: true
          description: "Optional human-readable label for the key"
    APIKeyCreated:
      type: object
      properties:
        id:
          type: string
        key:
          type: string
          description: "Raw API key — returned once only"
        prefix:
          type: string
          description: "First 8 characters of the key for identification"
        label:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    APIKeySummary:
      type: object
      properties:
        id:
          type: string
        prefix:
          type: string
        label:
          type: string
          nullable: true
        last_used_at:
          type: string
          format: date-time
          nullable: true
        revoked_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time

    # ── Wallet ────────────────────────────────────────────────────────────
...
    CreateTransferRequest:
      type: object
      required: ["from_wallet_id", "to_wallet_id", "asset", "amount"]
      properties:
        from_wallet_id:
          type: string
          format: uuid
        to_wallet_id:
          type: string
          format: uuid
        asset:
          type: string
          description: "Asset code to transfer (e.g. USDC, XLM)"
        amount:
          type: string
          description: "Positive decimal amount"
    TransferResponse:
      type: object
      properties:
        id:
          type: string
        tx_hash:
          type: string
          description: "Stellar transaction hash (empty until submitted)"
        type:
          type: string
          enum: ["transfer", "conversion", "funding"]
        status:
          type: string
          enum: ["pending", "submitted", "confirmed", "failed", "reconciliation_failed"]
        from_wallet_id:
          type: string
        to_wallet_id:
          type: string
        asset:
          type: string
        amount:
          type: string
          description: "Gross amount as decimal string"
        fee_amount:
          type: string
          description: "Fee deducted as decimal string"
        net_amount:
          type: string
          description: "Amount after fee as decimal string"
        fee_bps:
          type: integer
          description: "Fee in basis points"
        created_at:
          type: string
          format: date-time

    # ── Batch ─────────────────────────────────────────────────────────────
...